airflow-etl/requirements.txt
apache-airflow
apache-airflow-providers-apache-spark
apache-airflow-providers-celery>=3.3.0
apache-airflow-providers-google>=10.1.0
apache-airflow-providers-dbt-cloud

sqlparse>=0.5.0,<0.6.0
airflow-dbt-python==2.2.0
dbt-core==1.8.0
dbt-bigquery==1.8.0

pandas
psycopg2==2.9.9
pytest
pyspark
requests
sqlalchemy
graphviz
geopy